Fixing waiver window open from Calendar
