+++ js/menus.js
@@ -3,7 +3,6 @@
 const { app, BrowserWindow, clipboard, dialog, shell } = require('electron');
 const { appConfig, getDetails } = require('./app-config.js');
 const { checkForUpdates } = require('./update-manager');
-const { getDateStr } = require('./date-aux.js');
 const {
     getSavedPreferences,
 } = require('./saved-preferences.js');
@@ -13,7 +12,7 @@ const { savePreferences } = require('./user-preferences.js');
 const path = require('path');
 const Store = require('electron-store');
 const i18n = require('../src/configs/i18next.config');
-let { waiverWindow, prefWindow } = require('./windows');
+let { openWaiverManagerWindow, prefWindow } = require('./windows');
 
 function getMainMenuTemplate(mainWindow)
 {
@@ -23,35 +22,7 @@ function getMainMenuTemplate(mainWindow)
             id: 'workday-waiver-manager',
             click(item, window, event)
             {
-                if (waiverWindow !== null)
-                {
-                    waiverWindow.show();
-                    return;
-                }
-
-                if (event)
-                {
-                    const today = new Date();
-                    global.waiverDay = getDateStr(today);
-                }
-                const htmlPath = path.join('file://', __dirname, '../src/workday-waiver.html');
-                waiverWindow = new BrowserWindow({ width: 600,
-                    height: 500,
-                    parent: mainWindow,
-                    resizable: true,
-                    icon: appConfig.iconpath,
-                    webPreferences: {
-                        enableRemoteModule: true,
-                        nodeIntegration: true
-                    } });
-                waiverWindow.setMenu(null);
-                waiverWindow.loadURL(htmlPath);
-                waiverWindow.show();
-                waiverWindow.on('close', function()
-                {
-                    waiverWindow = null;
-                    mainWindow.webContents.send('WAIVER_SAVED');
-                });
+                openWaiverManagerWindow(mainWindow, event);
             },
         },
         {type: 'separator'},
