+++ js/windows.js
@@ -1,5 +1,10 @@
 'use strict';
 
+const { BrowserWindow } = require('electron');
+const { appConfig } = require('./app-config.js');
+const path = require('path');
+const { getDateStr } = require('./date-aux.js');
+
 // Keep a global reference of the window object, if you don't, the window will
 // be closed automatically when the JavaScript object is garbage collected.
 let waiverWindow = null;
@@ -7,9 +12,43 @@ let prefWindow = null;
 let tray = null;
 let contextMenu = null;
 
+function openWaiverManagerWindow(mainWindow, event)
+{
+    if (waiverWindow !== null)
+    {
+        waiverWindow.show();
+        return;
+    }
+
+    if (event)
+    {
+        const today = new Date();
+        global.waiverDay = getDateStr(today);
+    }
+    const htmlPath = path.join('file://', __dirname, '../src/workday-waiver.html');
+    waiverWindow = new BrowserWindow({ width: 600,
+        height: 500,
+        parent: mainWindow,
+        resizable: true,
+        icon: appConfig.iconpath,
+        webPreferences: {
+            enableRemoteModule: true,
+            nodeIntegration: true
+        } });
+    waiverWindow.setMenu(null);
+    waiverWindow.loadURL(htmlPath);
+    waiverWindow.show();
+    waiverWindow.on('close', function()
+    {
+        waiverWindow = null;
+        mainWindow.webContents.send('WAIVER_SAVED');
+    });
+}
+
 module.exports = {
     waiverWindow,
     prefWindow,
     tray,
-    contextMenu
+    contextMenu,
+    openWaiverManagerWindow
 };
